Endpoint to get config_id given product_feature
