+++ agent/lm_agent/backend_utils.py
@@ -86,6 +86,12 @@ async def get_bookings_from_backend(cluster_name: Optional[str] = None) -> List[
     return bookings
 
 
+async def get_config_id_from_backend(product_feature: str) -> int:
+    path = "/api/v1/config/"
+    resp = await async_client().get(path, params={"product_feature": product_feature})
+    return resp.json()
+
+
 async def get_config_from_backend() -> List[BackendConfigurationRow]:
     client = async_client()
     path = GET_CONFIG_URL_PATH
